Revert "added context (#11)"
This reverts commit f13e303e8397b2e1ec43865c9830439f9b31b06b.
