+++ packages/react-app/package.json
@@ -89,10 +89,10 @@
     }
   },
   "scripts": {
-    "build": "react-scripts build --openssl-legacy-provider",
-    "start": "react-scripts start --openssl-legacy-provider",
+    "build": "react-scripts --openssl-legacy-provider build",
     "eject": "react-scripts eject",
     "prestart": "node ./scripts/create_contracts.js",
+    "start": "react-scripts --openssl-legacy-provider start",
     "test": "react-scripts test",
     "lint": "eslint ./src/**/*.{js,jsx} --ignore-path ./.eslintignore --max-warnings=0",
     "ipfs": "node ./scripts/ipfs.js",
